server: Start versioning the DB schema
In preparation for #67.
