+++ server/src/domain/mod.rs
@@ -2,5 +2,6 @@ pub mod error;
 pub mod handler;
 pub mod opaque_handler;
 pub mod sql_backend_handler;
+pub mod sql_migrations;
 pub mod sql_opaque_handler;
 pub mod sql_tables;
